+++ app/src/main/java/com/nbusy/app/ChatDetailFragment.java
@@ -26,7 +26,7 @@ public class ChatDetailFragment extends ListFragment implements View.OnClickList
     // 3) bind it to application context: http://stackoverflow.com/questions/15235773/bind-service-to-fragmentactivity-or-fragment
     // 4) ** get it with IOC and never bother with service aspects as we're all in same process and same thread and just use workerService.sendMsg(...)
 
-    private final Worker worker = WorkerSingleton.getWorker(); // todo: get this from activity during onCreate()
+    private final Worker worker = WorkerSingleton.getWorker();
     private String chatId;
 
     public ChatDetailFragment() {
